+++ server/src/app/controllers/ContactController.js
@@ -1,5 +1,7 @@
 const ContactsRepository = require('../repositories/ContactsRepository');
 
+const isValidUUID = require('../utils/isValidUUID');
+
 class ContactController {
   async index(request, response) {
     const { orderBy } = request.query;
@@ -9,6 +11,10 @@ class ContactController {
 
   async show(request, response) {
     const { id } = request.params;
+
+    if (!isValidUUID(id)) {
+      return response.status(400).json({ error: 'Invalid Contact ID' });
+    }
 
     const contact = await ContactsRepository.findById(id);
 
@@ -26,16 +32,23 @@ class ContactController {
       return response.status(400).json({ error: 'Name is required' });
     }
 
-    const contactExists = await ContactsRepository.findByEmail(email);
-    if (contactExists) {
-      return response.status(400).json({ error: 'This e-mail is already in use' });
+    if (category_id && !isValidUUID(category_id)) {
+      return response.status(400).json({ error: 'Invalid Category' });
+    }
+
+    if (email) {
+      const contactExists = await ContactsRepository.findByEmail(email);
+
+      if (contactExists) {
+        return response.status(400).json({ error: 'This e-mail is already in use' });
+      }
     }
 
     const contact = await ContactsRepository.create({
       name,
-      email,
+      email: email,
       phone,
-      category_id,
+      category_id: category_id || null,
     });
 
     return response.status(201).json(contact);
@@ -45,25 +58,37 @@ class ContactController {
     const { id } = request.params;
     const { name, email, phone, category_id } = request.body;
 
-    const contactExists = await ContactsRepository.findById(id);
-    if (!contactExists) {
-      return response.status(404).json({ error: 'Contact not found' });
+    if (!isValidUUID(id)) {
+      return response.status(400).json({ error: 'Invalid Contact ID' });
+    }
+
+    if (category_id && !isValidUUID(category_id)) {
+      return response.status(400).json({ error: 'Invalid Category' });
     }
 
     if (!name) {
       return response.status(400).json({ error: 'Name is required' });
     }
 
-    const contactByEmail = await ContactsRepository.findByEmail(email);
-    if (contactByEmail && contactByEmail.id !== id) {
-      return response.status(404).json({ error: 'This e-mail is already in use' });
+    const contactExists = await ContactsRepository.findById(id);
+
+    if (!contactExists) {
+      return response.status(404).json({ error: 'Contact not found' });
+    }
+
+    if (email) {
+      const contactByEmail = await ContactsRepository.findByEmail(email);
+
+      if (contactByEmail && contactByEmail.id !== id) {
+        return response.status(404).json({ error: 'This e-mail is already in use' });
+      }
     }
 
     const contact = await ContactsRepository.update(id, {
       name,
-      email,
+      email: email || null,
       phone,
-      category_id,
+      category_id: category_id || null,
     });
 
     return response.json(contact);
@@ -71,6 +96,10 @@ class ContactController {
 
   async delete(request, response) {
     const { id } = request.params;
+
+    if (!isValidUUID(id)) {
+      return response.status(400).json({ error: 'Invalid Contact ID' });
+    }
 
     await ContactsRepository.delete(id);
     return response.sendStatus(204);
